Created packages / restructured classes Fixed null pointer when searching for non existing Customer
