+++ src/test/java/com/example/accessingdatamongodb/ApplicationControllerTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class AccessingDataMongodbRestApplicationTests {
+class ApplicationControllerTests {
 
 	@Test
 	void contextLoads() {
